feat: modularize and reuse divider component
