+++ app/document/[documentId]/page.module.css
@@ -2,17 +2,7 @@
   width: 100%;
   height: 100%;
   background-color: var(--secondary-dark);
-}
-
-.divider {
-  background-color: var(--primary-light);
-  width: 15px;
-  cursor: col-resize;
-  user-select: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--background);
+  overflow-y: auto;
 }
 
 .icon {
@@ -83,7 +73,6 @@
 
 .alertContainer {
   transition: transform 0.3s ease-out;
-  opacity: 0.9;
   flex-direction: row;
   margin-top: 10px;
   z-index: 5;
@@ -91,8 +80,8 @@
   align-items: center;
   color: var(--primary-dark);
   position: fixed;
-  background-color: var(--background);
-  border: 1px solid #c5c5c583;
+  background-color: var(--background-accent);
+  border: 1px solid var(--grey-normal);
   border-radius: 10px 0 0 10px;
   top: 0;
   right: 0;
